Fix bottle position code

+++ firmware/node/defs.h
@@ -17,7 +17,7 @@ typedef struct
 
 extern uint8_t g_error;
 extern uint8_t g_node_id;
-extern int32_t g_pos[3];
+extern int16_t g_pos[3];
 extern uint32_t g_random_seed;
 
 void print_col(color_t *col);
